Fixed wrong CPU mapping
The reason of the fix is related to the fact that `cpu_sub_mapping` maps
from a reduced interval of indices of cores to indices of cores. However,
`common->cpu_info` stores only a reduced number of core indices in case
of disabled hyperthreading. The previous behaviour lead to picking
uninitialized `cpu_info`s, this should fix it.